backend/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sophikon-backend"
version = "0.1.0"
description = "Sophikon V1.0 - AI-powered project management backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core Framework
    "fastapi>=0.128.2",
    "uvicorn[standard]>=0.40.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.46",
    "asyncpg>=0.31.0",
    "psycopg2-binary>=2.9.0", # For Alembic migrations
    "alembic>=1.18.3",
    # Authentication
    "python-jose[cryptography]>=3.5.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.22",
    # Rate Limiting
    "slowapi>=0.1.9",
    # Redis
    "redis>=7.1.0",
    # Background Tasks
    "celery>=5.6.2",
    # Email
    "fastapi-mail>=1.6.1",
    # File Handling
    "pillow>=12.1.0",
    # Utilities
    "uuid-utils>=0.14.0",
    "email-validator>=2.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.28.1",  # For testing
    "faker>=33.3.0",  # Test data generation
    "ruff>=0.9.0",    # Linting
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=app --cov-report=term-missing"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
pythonpath = [
    ".",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["alembic*", "tests*"]


[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]
